Restructure the folder structure to organize the code properly.
