+++ src/App.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { BrowserRouter, Route } from 'react-router-dom';
-import NavigationBar from './NavigationBar';
-import Home from './Home';
-import SignInContainer from '../containers/SignInContainer';
-import RegistrationsContainer from '../containers/RegistrationsContainer';
+import NavigationBar from './shared/components/NavigationBar';
+import Home from './home/components/Home';
+import SignInContainer from './users/containers/SignInContainer';
+import RegistrationsContainer from './users/containers/RegistrationsContainer';
 
 const App = () => (
   <BrowserRouter>
