+++ src/shared/actionTypes.js
