+++ src/shared/components/NavigationBar.js
